Recode Warp Prediction
Instead of using my bootleg warp prediction which relied on velocity to predict where the target's next position would be. We instead use the server to get the most recent position of our target, we can compensate for the choked commands the player did, usually by getting the delta of OldSimTime you are able to add that delta into the players simulation time in order to move him into place on the server, but you still need a position you can shoot at and hit because backtracking is impossible when lagcompensated, so we instead predict the movement of said target so that it aligns with the servers position and you shoot at THAT position while also correcting aimbot. There you end up with a very general idea of where our target may be, though I am uncertain if I did this correctly or if I am missing a few things in my own version.